requirements.txt
kafka-python-ng
pandas
sqlalchemy
streamlit
watchdog
